Text.Wrap

Documentation

data WrapSettings

Settings to control how wrapping is performed.

Constructors

WrapSettings

Fields

preserveIndentation :: Bool

Whether to indent new lines created by wrapping when their original line was indented.

breakLongWords :: Bool

Whether to break in the middle of the first word on a line when that word exceeds the wrapping width.

defaultWrapSettings :: WrapSettings

wrapTextToLines :: WrapSettings -> Int -> Text -> [Text]

Wrap text at the specified width. Newlines and whitespace in the input text are preserved. Returns the lines of text in wrapped form. New lines introduced due to wrapping will have leading whitespace stripped.

wrapText :: WrapSettings -> Int -> Text -> Text

Like wrapTextToLines, but returns the wrapped text reconstructed with newlines inserted at wrap points.